Add command component: Msg(Target|To)
An enum allowing a broad variety of targets for messages and commands,
and a vector of multiple targets.

The goal is to standardize message parsing by using enums and structs to
validate and represent protocol-level types. These types are not yet
wired up; the immediate goal is to implement the syntax requirements of
RFC 2812.

https://tools.ietf.org/html/rfc2812#section-2.3.1